[refactor] 방만들기 페이지 폴더 정리 - pull로 당겨온 후 파일 정리
