+++ src/pages/CreateRoom.jsx
@@ -3,17 +3,17 @@ import { useRef, useEffect, useState, useContext } from 'react';
 import { pb } from "@/api/pocketbase";
 import FormInput from "@/components/FormInput";
 import Button from '@/components/Button';
-import ContentTextarea from '@/parts/creaate/ContentTextarea';
-import DatePicker from "@/parts/creaate/DatePicker";
+import ContentTextarea from '@/parts/create/ContentTextarea';
+import DatePicker from "@/parts/create/DatePicker";
 // import TimePicker from "@/parts/creaate/TimePicker";
-import ParticipateCounter from "@/parts/creaate/ParticipateCounter";
-import MeetingPoint from "@/parts/creaate/MeetingPoint";
+import ParticipateCounter from "@/parts/create/ParticipateCounter";
+import MeetingPoint from "@/parts/create/MeetingPoint";
 // import Location from "@/parts/map/Location";
-import CategoryDropdown from "@/parts/creaate/CategoryDropdown";
-import Status from '@/parts/creaate/Status';
-import PaymentToggleButton from '@/parts/creaate/PaymentToggleButton';
+import CategoryDropdown from "@/parts/create/CategoryDropdown";
+import Status from '@/parts/create/Status';
+import PaymentToggleButton from '@/parts/create/PaymentToggleButton';
 // import { useNavigate } from 'react-router-dom';
-import FileUpload from '../parts/creaate/FileUpload';
+import FileUpload from '../parts/create/FileUpload';
 import CreateHeader from '@/layout/CreateHeader';
 import { AppContext } from '@/App';
