--- src/parts/creaate/DatePicker.jsx
@@ -1,17 +1,0 @@
-
-import { useState, forwardRef, useId, } from "react";
-// import Input from "@/components/Input";
-
-function DatePicker({ title, className, labelClassName, ...restProps }, ref) {
-  const [date, setDate] = useState(null);
-  const id = useId();
-
-  return (
-    <div>
-      <label htmlFor={id} className={labelClassName}>{title}</label>
-      <input id={id} ref={ref} className={className} type="date" onChange={(e) => setDate(e.target.value)} {...restProps} />
-    </div>
-  );
-};
-
-export default forwardRef(DatePicker);
